move public key config vars from server to common
