+++ opal_common/schemas/store.py
@@ -1,4 +1,4 @@
-from typing import List, Dict, Any, Optional
+from typing import List, Dict, Any
 from pydantic import BaseModel, Field
 
 class StoreTransaction(BaseModel):
