+++ opal_server/server.py
@@ -1,5 +1,3 @@
-from opal_server.security.api import init_security_router
-from opal_server.security.jwks import JwksStaticEndpoint
 import os
 import asyncio
 from functools import partial
@@ -14,10 +12,13 @@ from opal_common.schemas.data import ServerDataSourceConfig
 from opal_common.synchronization.named_lock import NamedLock
 from opal_common.middleware import configure_middleware
 from opal_common.authentication.signer import JWTSigner
+from opal_common.authentication.deps import JWTAuthenticator, StaticBearerAuthenticator
+from opal_common.config import opal_common_config
 from opal_server.config import opal_server_config
+from opal_server.security.api import init_security_router
+from opal_server.security.jwks import JwksStaticEndpoint
 from opal_server.data.api import init_data_updates_router
 from opal_server.data.data_update_publisher import DataUpdatePublisher
-from opal_common.authentication.deps import JWTAuthenticator, StaticBearerAuthenticator
 from opal_server.policy.bundles.api import router as bundles_router
 from opal_server.policy.github_webhook.api import init_git_webhook_router
 from opal_server.policy.watcher import (setup_watcher_task,
@@ -91,10 +92,10 @@ class OpalServer:
         else:
             self.signer = JWTSigner(
                 private_key=opal_server_config.AUTH_PRIVATE_KEY,
-                public_key=opal_server_config.AUTH_PUBLIC_KEY,
-                algorithm=opal_server_config.AUTH_JWT_ALGORITHM,
-                audience=opal_server_config.AUTH_JWT_AUDIENCE,
-                issuer=opal_server_config.AUTH_JWT_ISSUER,
+                public_key=opal_common_config.AUTH_PUBLIC_KEY,
+                algorithm=opal_common_config.AUTH_JWT_ALGORITHM,
+                audience=opal_common_config.AUTH_JWT_AUDIENCE,
+                issuer=opal_common_config.AUTH_JWT_ISSUER,
             )
 
         if enable_jwks_endpoint:
